Add a new Search API to Onyx based on SearchTool
At Stacklok, we work on an OSS project called ToolHive, which is a framework for developing MCP servers. Using ToolHive, we implemented a new MCP server for Onyx. We are piloting this MCP server internally at Stacklok. The MCP server calls the new Onyx search API. Through this PR, we would like to contribute the search API code to the Onyx project. We plan to open-source the MCP server soon.

We developed the search API based on our (limited) understanding of Onyx application design. We would like to get your feedback on its implementation.

Below are the detailed parameters of the search API:
- query: User’s search query
- time_cutoff: Search for documents created on or after this date
- document_sources: Search for documents from these sources

The endpoint is protected with an API KEY that needs to be generated from the Onyx UI.

Usage:
```bash
curl -X POST "http://localhost:8080/onyx-tools/search-tool" \
    -H "Content-Type: application/json" \
    -H "Authorization: Bearer $ONYX_API_KEY" \
    -d '{
        "query": "key projects stacklok",
        "time_cutoff": "2024-01-01T00:00:00Z",
        "document_sources": ["google_drive"]
    }' | jq
```