feat(auth): Add Auth0 support for authentication.
This commit introduces Auth0 as a primary authentication method for Onyx, providing a robust and secure way for users to log in via Single Sign-On (SSO). This change is particularly important for enterprise customers who need to integrate Onyx with their existing identity providers like Microsoft Active Directory.

This feature has been thoroughly tested in our production environment, including a successful implementation with Microsoft Active Directory integrated through Auth0.

We implemented the following three options to be included in the .env:
- AUTH0_CLIENT_ID=
- AUTH0_CLIENT_SECRET=
- AUTH0_DOMAIN=
In addition to this, you also have to set the authentication type to
Auth0.
- AUTH_TYPE=auth0